pyproject.toml
[tool.poetry]
name = "projeto-academia-churn"
version = "0.1.0"
description = "API com FastAPI e modelo de previsão de churn"
authors = ["yagovinicius <yagoviniciusribeiro05@gmail.com>"]
packages = [
    { include = "app" },
    { include = "ml" }
]

[tool.poetry.dependencies]
python = ">=3.13,<4.0"
pandas = ">=2.3.0,<3.0.0"
fastapi = ">=0.115.1,<0.116.0"
scikit-learn = ">=1.7.0,<2.0.0"
uvicorn = { version = ">=0.34.3,<0.35.0", extras = ["standard"] }
psycopg2-binary = ">=2.9.10,<3.0.0"
python-dotenv = ">=1.1.0,<2.0.0"
sqlalchemy = ">=2.0.41,<3.0.0"
pydantic = { version = ">=2.11.7,<3.0.0", extras = ["email"] }
alembic = ">=1.16.2,<2.0.0"
passlib = { version = ">=1.7.4,<2.0.0", extras = ["bcrypt"] }
joblib = ">=1.4.2,<2.0.0"
matplotlib = ">=3.9.0,<4.0.0"
seaborn = "^0.13.2"
openpyxl = "^3.1.5"
xgboost = "^3.0.2"
pika = "^1.3.2"
apscheduler = "^3.11.0"
pytest = "^8.4.1"
httpx = "^0.28.1"
python-jose = "^3.5.0"
python-multipart = "^0.0.20"
faker = "^37.4.0"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
